Fixed Findbugs issues
